+++ types/ts/app/interfaces/common.d.ts
@@ -28,6 +28,7 @@ export interface ImageParams {
     inlineStyles?: string;
     captionTemplate?: string;
     captionCssClass?: string;
+    captionInlineStyles?: string;
 }
 export interface ButtonParams {
     cssClass?: string;